+++ blocksuite/affine/data-view/src/core/group-by/setting.ts
@@ -17,7 +17,7 @@ import { css, html, unsafeCSS } from 'lit';
 import { property, query } from 'lit/decorators.js';
 import { repeat } from 'lit/directives/repeat.js';
 
-import { isKanbanGroupableProperty } from '../../view-presets/kanban/group-by-utils.js';
+import { canGroupable } from '../../view-presets/kanban/group-by-utils.js';
 import { KanbanSingleView } from '../../view-presets/kanban/kanban-view-manager.js';
 import { TableSingleView } from '../../view-presets/table/table-view-manager.js';
 import { dataViewCssVariable } from '../common/css-variable.js';
@@ -280,10 +280,7 @@ export const selectGroupByProperty = (
               return false;
             }
             if (view instanceof KanbanSingleView) {
-              return isKanbanGroupableProperty(
-                view.manager.dataSource,
-                property.id
-              );
+              return canGroupable(view.manager.dataSource, property.id);
             }
             const dataType = property.dataType$.value;
             if (!dataType) {
